[CWS] fix testsuite duplicated init functions
